Separated library into modules, added documentation
